refactor: separate localeContext from Layout component
