+++ src/components/localeContext.js
@@ -1,0 +1,5 @@
+import React from 'react';
+
+const LocaleContext = React.createContext();
+
+export default LocaleContext;
