+++ src/components/useTranslations.tsx
@@ -1,12 +1,12 @@
 import React from 'react';
 import { useStaticQuery, graphql } from 'gatsby';
-import { LocaleContext } from './layout';
+import LocaleContext from './localeContext';
 
 interface ITranslateItem {
   node: {
     name: string;
     translations: string;
-  }
+  };
 }
 
 interface ILang {
@@ -16,12 +16,12 @@ interface ILang {
 function useTranslations() {
   const { locale } = React.useContext(LocaleContext);
   const { rawData } = useStaticQuery(query);
-  const { translations } = rawData.edges.map((item: ITranslateItem) => (
-    {
+  const { translations } = rawData.edges
+    .map((item: ITranslateItem) => ({
       name: item.node.name,
-      translations: item.node.translations
-    }
-  )).find((lang: ILang) => lang.name === locale);
+      translations: item.node.translations,
+    }))
+    .find((lang: ILang) => lang.name === locale);
   return translations;
 }
 
@@ -29,9 +29,7 @@ export default useTranslations;
 
 const query = graphql`
   query useTranslations {
-    rawData: allFile(
-      filter: { sourceInstanceName: { eq: "translations" } }
-    ) {
+    rawData: allFile(filter: { sourceInstanceName: { eq: "translations" } }) {
       edges {
         node {
           name
