+++ src/styles/index.scss
@@ -42,7 +42,6 @@
 
 .main-content {
   flex-grow: 1;
-  height: 150vh;
 }
 
 @media (min-width: 576px) {
