+++ wrapPageElement.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import { Layout } from './src/components/layout';
+import Layout from './src/components/layout';
 
 import 'bootstrap/dist/css/bootstrap.min.css';
 import './src/styles/index.scss';
